refactor(rust): move impls out of model, scalars
Just a bunch of cleanups
