+++ rust/mlt-core/src/frames/v01/geometry/model.rs
@@ -77,27 +77,6 @@ pub enum GeometryType {
     MultiPolygon,
 }
 
-impl GeometryType {
-    #[must_use]
-    pub fn is_polygon(self) -> bool {
-        matches!(self, GeometryType::Polygon | GeometryType::MultiPolygon)
-    }
-    #[must_use]
-    pub fn is_linestring(self) -> bool {
-        matches!(
-            self,
-            GeometryType::LineString | GeometryType::MultiLineString
-        )
-    }
-    #[must_use]
-    pub fn is_multi(self) -> bool {
-        matches!(
-            self,
-            GeometryType::MultiPoint | GeometryType::MultiLineString | GeometryType::MultiPolygon
-        )
-    }
-}
-
 /// Describes how the vertex buffer should be encoded.
 #[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Default)]
 #[cfg_attr(test, derive(proptest_derive::Arbitrary))]
